requirements/development.txt
APScheduler==3.6.0
backports.functools-lru-cache==1.6.1
billiard==4.1.0
boto3==1.28.27
botocore==1.31.27
celery==5.3.1
certifi==2023.7.22
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.12.2
Django==4.2.4
django-bulk-update==2.2.0
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
django-enumfields==2.1.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.2
django-nine==0.2.7
django-ratelimit==2.0.0
django-redis==5.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
docutils==0.14
drf-enum-field==0.9.2
elastic-apm==6.11.0
elasticsearch==7.17.6
elasticsearch-dsl>=7.2.0<8.0.0
hexbytes==0.2.1
idna==2.7
importlib-metadata==3.3.0
Jinja2==3.1.2
jmespath==0.9.3
kafka-python==2.0.1
kombu==5.3.1
lazy-object-proxy==1.4.3
lru-dict==1.1.6
Markdown==3.3.3
MarkupSafe==2.1.3
parsimonious==0.9.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pybase62==0.3.3
pycparser==2.20
python-dateutil==2.8.2
python3-logstash==0.4.80
pytz==2022.1
redis==5.0.0
requests==2.31.0
rlp==3.0.0
s3transfer==0.6.0
simplejson==3.16.0
six==1.16.0
sqlparse==0.4.4
toolz==0.11.1
typed-ast==1.5.4
tzlocal==2.0
uritemplate==3.0.0
urllib3==1.26.9
web3==6.8.0
websockets==10.1
pika==1.3.2
Pillow==10.0.0
google-cloud-secret-manager
google-cloud-storage